requirements.txt
anthropic>=0.20.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=7.4.0
pytest-cov>=4.0.0
